pyproject.toml
[project]
name = "aadk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "db-dtypes>=1.5.0",
    "google-adk>=1.21.0",
    "langchain>=1.2.0",
    "langchain-core>=1.2.5",
    "langchain-google-genai>=4.1.2",
    "pandas>=2.3.3",
    "pypdf>=6.5.0",
]
